fixed removal of exemptions
